+++ R/cre.R
@@ -49,9 +49,7 @@
 #'  - *max_depth*: Maximum rules length (default: 3).
 #'  - *replace*: Boolean variable for replacement in bootstrapping
 #'  (default: TRUE).
-#'  - *max_decay*: Decay Threshold for pruning the rules (default: 0.025).
-#'  - *type_decay*: Decay Type for pruning the rules: 1 relative error; 2 error
-#'  (default: 2).
+#'  - *t_decay*: The decay threshold for rules pruning (default: 0.025).
 #'  - *t_ext*: The threshold to define too generic or too specific (extreme)
 #'  rules (default: 0.01, range: (0,0.5)).
 #'  - *t_corr*: The threshold to define correlated rules (default: 1,
@@ -108,8 +106,7 @@
 #'                      node_size = 20,
 #'                      max_nodes = 5,
 #'                      max_depth = 3,
-#'                      max_decay = 0.025,
-#'                      type_decay = 2,
+#'                      t_decay = 0.025,
 #'                      t_ext = 0.025,
 #'                      t_corr = 1,
 #'                      t_pvalue = 0.05,