+++ tests/testthat/test-discover_rules.R
@@ -17,20 +17,15 @@ test_that("discover_rules works as expected!", {
                         ps_method_inf = "SL.xgboost",
                         oreg_method_inf = "SL.xgboost",
                         include_ps_inf = TRUE,
-                        include_offset = FALSE,
-                        cate_method = "DRLearner",
-                        cate_SL_library = "SL.xgboost",
-                        offset_name = NA,
-                        random_state = 3591)
+                        offset = NA)
 
   hyper_params <- list(intervention_vars = c(),
                        ntrees_rf = 100,
                        ntrees_gbm = 50,
                        node_size = 20,
                        max_nodes = 5,
-                       max_depth = 15,
-                       max_decay = 0.025,
-                       type_decay = 2,
+                       max_depth = 3,
+                       t_decay = 0.025,
                        t_ext = 0.025,
                        t_corr = 1,
                        t_pvalue = 0.05,
@@ -54,8 +49,7 @@ test_that("discover_rules works as expected!", {
                 ps_method = getElement(method_params, "ps_method_dis"),
                 oreg_method = getElement(method_params, "oreg_method_dis"),
                 X_names = X_names,
-                include_offset = getElement(method_params, "include_offset"),
-                offset_name = getElement(method_params, "offset_name"),
+                offset = getElement(method_params, "offset"),
                 random_state = getElement(method_params, "random_state"))
   ite <- ite_list[["ite"]]