+++ tests/testthat/test-filter_correlated_rules.R
@@ -17,8 +17,7 @@ test_that("Correlated Rules Discarded Correctly", {
   max_nodes <- 5
   max_depth <- 15
   replace <- TRUE
-  max_decay <- 0.025
-  type_decay <- 2
+  t_decay <- 0.025
   t_corr <- 1
   t_ext <- 0.01
   intervention_vars <- c()
@@ -47,7 +46,7 @@ test_that("Correlated Rules Discarded Correctly", {
                                   replace)
 
   rules_list <- filter_irrelevant_rules(initial_rules, X,
-                                        ite, max_decay, type_decay)
+                                        ite, t_decay)
   rules_matrix <- generate_rules_matrix(X, rules_list)
 
   rules_matrix <- filter_extreme_rules(rules_matrix, rules_list, t_ext)