+++ tests/testthat/test-filter_extreme_rules.R
@@ -17,8 +17,7 @@ test_that("Extreme Rules Discarded Correctly", {
   max_nodes <- 5
   max_depth <- 15
   replace <- FALSE
-  max_decay <- 0.025
-  type_decay <- 2
+  t_decay <- 0.025
   t_ext <- 0.1
   intervention_vars <- c()
 
@@ -45,8 +44,7 @@ test_that("Extreme Rules Discarded Correctly", {
                                   ntrees_gbm, node_size, max_nodes, max_depth,
                                   replace)
 
-  rules_list <- filter_irrelevant_rules(initial_rules, X, ite,
-                                        max_decay, type_decay)
+  rules_list <- filter_irrelevant_rules(initial_rules, X, ite, t_decay)
   rules_matrix <- generate_rules_matrix(X, rules_list)
 
   ###### Run Tests ######