+++ tests/testthat/test-generate_rules_matrix.R
@@ -17,8 +17,7 @@ test_that("Rules Extracted Correctly", {
   max_nodes <- 5
   max_depth <- 15
   replace <- FALSE
-  max_decay <- 0.025
-  type_decay <- 2
+  t_decay <- 0.025
   intervention_vars <- c()
 
   # Check for binary outcome
@@ -44,7 +43,7 @@ test_that("Rules Extracted Correctly", {
                                   ntrees_gbm, node_size, max_nodes, max_depth,
                                   replace)
 
-  rules <- filter_irrelevant_rules(initial_rules, X, ite, max_decay, type_decay)
+  rules <- filter_irrelevant_rules(initial_rules, X, ite, t_decay)
 
 
   ###### Run Tests ######